pythonEnhance/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
Flask==2.2.5
fsspec==2023.1.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
PyYAML==6.0
requests==2.31.0
safetensors==0.3.1
timm==0.9.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
